Move command error handling traits
